Major refactoring of structure of accounts service
