+++ src/accounts/app/main.py
@@ -1,0 +1,5 @@
+import app.server.Accounts as account_service
+import app.database.DatabaseController as db
+if __name__ == "__main__":
+
+    account_service.start_service()
