+++ src/accounts/tests/UserService_unit_test.py
@@ -4,16 +4,16 @@ from app.database.DatabaseController import DatabaseController
 from app.authentication.AuthenticationManager import AuthenticationManager
 from app.database.Mockdata import Mockdata
 
-class TestAuthenticationManager(unittest.TestCase):
+class TestUserService(unittest.TestCase):
     
     def setUp(self):
         self.database = Mockdata()
         self.database_controller = DatabaseController(self.database)
         self.am = AuthenticationManager(self.database_controller)
 
-    def test_get_hash(self):
+    def test_create_new_guest(self):
         pass
-    def test_check_hash(self):
+    def test_create_new_staff(self):
         pass
-    def test_authenticate_staff_login(self):
+    def test_delete_user(self):
         pass
